+++ backend/src/compiler/expr/logic.rs
@@ -6,7 +6,7 @@ use aelys_syntax::ast::Expr;
 
 impl Compiler {
     // Short-circuit: if left is false, skip right
-    pub fn compile_and(&mut self, left: &Expr, right: &Expr, dest: u8, span: Span) -> Result<()> {
+    pub fn compile_and(&mut self, left: &Expr, right: &Expr, dest: u16, span: Span) -> Result<()> {
         self.compile_expr(left, dest)?;
         let jump = self.emit_jump_if(OpCode::JumpIfNot, dest, span);
         self.compile_expr(right, dest)?;
@@ -15,7 +15,7 @@ impl Compiler {
     }
 
     // Short-circuit: if left is true, skip right
-    pub fn compile_or(&mut self, left: &Expr, right: &Expr, dest: u8, span: Span) -> Result<()> {
+    pub fn compile_or(&mut self, left: &Expr, right: &Expr, dest: u16, span: Span) -> Result<()> {
         self.compile_expr(left, dest)?;
         let jump = self.emit_jump_if(OpCode::JumpIf, dest, span);
         self.compile_expr(right, dest)?;
