+++ runtime/src/vm/call_data.rs
@@ -4,8 +4,8 @@ use super::{GcRef, NativeFunction, Value};
 pub(super) enum CallData {
     Function {
         func_ref: GcRef,
-        arity: u8,
-        num_registers: u8,
+        arity: u16,
+        num_registers: u32,
         bytecode_ptr: *const u32,
         bytecode_len: usize,
         constants_ptr: *const Value,
@@ -16,8 +16,8 @@ pub(super) enum CallData {
     },
     Closure {
         inner_func_ref: GcRef,
-        arity: u8,
-        num_registers: u8,
+        arity: u16,
+        num_registers: u32,
         bytecode_ptr: *const u32,
         bytecode_len: usize,
         constants_ptr: *const Value,
